+++ src/main.cpp
@@ -25,11 +25,11 @@ double stop_timer() {
 	return double(chrono::duration_cast<chrono::nanoseconds>(end_time - start_time).count());
 }
 
-tuple<vector<vector<int> >, vector<Estacion> > cargar_input(int n, int m);
+tuple<vector<vector<double> >, vector<Estacion> > cargar_input(int n, int m);
 
-int distTrigonometrica(tuple<int, int, int> a, tuple<int, int, int> b);
+double distTrigonometrica(tuple<int, int, int> a, tuple<int, int, int> b);
 
-void imprimir_res(tuple<int,int,vector<int > > res);
+void imprimir_res(tuple<double,int,vector<int > > res);
 
 int main(int argc, char *argv[]) {
 	int numeroDeEjercicio = 0;
@@ -57,10 +57,10 @@ int main(int argc, char *argv[]) {
     cout << "ingrese en las siguientes " << n << " lineas xg, yg, pg, ubicacion y cantidad de posiones necesarias de los gym" << endl;
     cout << "luego, en las siguientes " << m << " lineas xp, yp ubicacion de las paradas" << endl;
 
-    tuple<vector<vector<int> >, vector<Estacion> > input = cargar_input(n,m);
-    vector<vector<int> > distancias = get<0>(input);
+    tuple<vector<vector<double> >, vector<Estacion> > input = cargar_input(n,m);
+    vector<vector<double> > distancias = get<0>(input);
     vector<Estacion> estaciones = get<1>(input);
-    std::tuple<int, int, std::vector<int> > res = solverEj1(estaciones, distancias, n, m, k);
+    std::tuple<double, int, std::vector<int> > res = solverEj1(estaciones, distancias, n, m, k);
     imprimir_res(res);
   }
   else if (numeroDeEjercicio == 2) {
@@ -70,10 +70,10 @@ int main(int argc, char *argv[]) {
     cout << "ingrese en las siguientes " << n << " lineas xg, yg, pg, ubicacion y cantidad de posiones necesarias de los gym" << endl;
     cout << "luego, en las siguientes " << m << " lineas xp, yp ubicacion de las paradas" << endl;
 
-    tuple<vector<vector<int> >, vector<Estacion> > input = cargar_input(n,m);
-    vector<vector<int> > distancias = get<0>(input);
+    tuple<vector<vector<double> >, vector<Estacion> > input = cargar_input(n,m);
+    vector<vector<double> > distancias = get<0>(input);
     vector<Estacion> estaciones = get<1>(input);
-    std::tuple<int, int, std::vector<int> > res = solverEj2(estaciones, distancias, n, m, k);
+    std::tuple<double, int, std::vector<int> > res = solverEj2(estaciones, distancias, n, m, k);
     imprimir_res(res);
   }
 
@@ -82,7 +82,7 @@ int main(int argc, char *argv[]) {
   return 0;
 }
 
-tuple<vector<vector<int> >, vector<Estacion> > cargar_input(int n, int m){
+tuple<vector<vector<double> >, vector<Estacion> > cargar_input(int n, int m){
   vector< tuple <int, int, int> > gimnasios_y_paradas;
 
   for (int i = 0; i < n; i++) {
@@ -100,10 +100,10 @@ tuple<vector<vector<int> >, vector<Estacion> > cargar_input(int n, int m){
     gimnasios_y_paradas.push_back(pp);
   }
 
-  vector<vector<int> > distancias(n+m);
+  vector<vector<double> > distancias(n+m);
   for (int i = 0; i < n+m; ++i) {
     for (int k = 0; k < n+m; ++k) {
-      int dist_ik = distTrigonometrica(gimnasios_y_paradas[i], gimnasios_y_paradas[k]);
+      double dist_ik = distTrigonometrica(gimnasios_y_paradas[i], gimnasios_y_paradas[k]);
       distancias[i].push_back(dist_ik);
     }
   }
@@ -117,17 +117,17 @@ tuple<vector<vector<int> >, vector<Estacion> > cargar_input(int n, int m){
 }
 
 
-int distTrigonometrica(tuple<int, int, int> a, tuple<int, int, int> b){
+double distTrigonometrica(tuple<int, int, int> a, tuple<int, int, int> b){
   int xa = get<0>(a);
   int ya = get<1>(a);
   int xb = get<0>(b);
   int yb = get<1>(b);
   int x = (abs(xa - xb))*(abs(xa - xb));
   int y = (abs(ya - yb))*(abs(ya - yb));
-  return (int) sqrt(x+y);
+  return sqrt(x+y);
 }
 
-void imprimir_res(tuple<int,int,vector<int > > res){
+void imprimir_res(tuple<double,int,vector<int > > res){
   cout << "D= " << get<0>(res) << " k= " << get<1>(res) << " i= ";
     for (int i = 0; i < get<1>(res); ++i) {
       cout << get<2>(res)[i]+1 << " ";
